fetch_price/requirements.txt
functions-framework
httpx
google-cloud-bigquery